+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- `Tuple6` through `Tuple8`
 
 ## [2.0.0] - 2017-11-13
 ### Changed